update result of state change refactor code
