+++ lib/core/routing/app_router.dart
@@ -1,6 +1,7 @@
 import 'package:go_router/go_router.dart';
 import 'package:warehouse_app/core/routing/routes.dart';
 import 'package:warehouse_app/features/Auth/presentation/views/auth_view.dart';
+import 'package:warehouse_app/features/home/presentation/views/home_view.dart';
 import 'package:warehouse_app/features/onboarding/presentation/views/onboarding_view.dart';
 
 abstract class AppRouter {
@@ -13,6 +14,10 @@ abstract class AppRouter {
       GoRoute(
         path: Routes.auth,
         builder: (context, state) => const AuthView(),
+      ),
+      GoRoute(
+        path: Routes.home,
+        builder: (context, state) => const HomeView(),
       ),
     ],
   );
