+++ lib/core/routing/routes.dart
@@ -1,9 +1,6 @@
 abstract class Routes {
   static const onboarding = '/';
   static const auth = '/auth';
-  static const register = '/register';
-  static const profile = '/profile';
-  static const settings = '/settings';
-  static const about = '/about';
-  static const notFound = '/404';
+  static const home = '/home';
+
 }
