+++ lib/features/Auth/presentation/views/widgets/auth_view_body.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:warehouse_app/core/utils/service_locator.dart';
+import 'package:warehouse_app/core/DI/service_locator.dart';
 import 'package:warehouse_app/features/Auth/presentation/cubits/auth_scroll_cubit/auth_scroll_cubit.dart';
 import 'package:warehouse_app/features/Auth/presentation/cubits/sign_up_cubit/sign_up_cubit.dart';
 import '../../../../../core/utils/assets.dart';
-import '../../../../../core/utils/spacer.dart';
+import '../../../../../core/helpers/spacer.dart';
 import '../../../../../core/utils/styles.dart';
 import 'container_of_auth_type_widget.dart';
 import 'login_view_widget.dart';
