+++ lib/features/Auth/presentation/views/widgets/sign_up_create_account_button_widget.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
+import 'package:warehouse_app/core/routing/routes.dart';
+import '../../../../../core/helpers/custom_snack_bar_method.dart';
 import '../../../../../core/widgets/app_text_button.dart';
 import '../../../../../generated/l10n.dart';
 import '../../cubits/sign_up_cubit/sign_up_cubit.dart';
@@ -15,28 +18,26 @@ class SignUpCreateAccountButtonWidget extends StatelessWidget {
     return BlocConsumer<SignUpCubit, SignUpState>(
       listener: (context, state) {
         if (state is SignUpFailure) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              content: Text(
-                state.errMes,
-              ),
-            ),
-          );
-        } else {
-          print('Success State ===>>>>>>>>>');
+          customSnackBarMethod(context, state.errMes);
+        } else if (state is SignUpSuccess) {
+          context.pushReplacement(Routes.home);
         }
       },
       builder: (context, state) {
         return AppTextButton(
           text: S.of(context).signUp,
           child: state is SignUpLoading
-              ? const Center(
-                  child: CircularProgressIndicator(),
+              ? Center(
+                  child: SizedBox(
+                    width: 24.r,
+                    height: 24.r,
+                    child: const CircularProgressIndicator(
+                      color: Colors.white,
+                    ),
+                  ),
                 )
               : null,
-          onPressed: () {
-            context.read<SignUpCubit>().signUp();
-          },
+          onPressed: () => context.read<SignUpCubit>().signUp(),
         );
       },
     );
