+++ lib/features/Auth/presentation/views/widgets/sign_up_data_fields.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../../core/utils/color_manager.dart';
-import '../../../../../core/utils/spacer.dart';
+import '../../../../../core/helpers/spacer.dart';
 import '../../../../../core/utils/styles.dart';
 import '../../../../../core/widgets/app_text_form_field.dart';
 import '../../../../../generated/l10n.dart';
