+++ lib/features/home/presentation/views/home_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold();
+  }
+}
